:sparkles: feat : query type update

+++ src/article/application/get_articles_default_usecase.rs
@@ -14,13 +14,11 @@ pub struct ListArticleRequest {
     tag: Option<String>,
     author: Option<String>,
     favorited: Option<String>,
-    limit: Option<usize>,
-    offset: Option<usize>,
+    limit: Option<i64>,
+    offset: Option<i64>,
 }
 
 impl ListArticleRequest {
-
-
 
     pub fn tag(&self) -> &Option<String> {
         &self.tag
@@ -31,13 +29,13 @@ impl ListArticleRequest {
     pub fn favorited(&self) -> &Option<String> {
         &self.favorited
     }
-    pub fn limit(&self) -> Option<usize> {
+    pub fn limit(&self) -> Option<i64> {
         self.limit
     }
-    pub fn offset(&self) -> Option<usize> {
+    pub fn offset(&self) -> Option<i64> {
         self.offset
     }
-    pub fn new(tag: Option<String>, author: Option<String>, favorited: Option<String>, limit: Option<usize>, offset: Option<usize>) -> Self {
+    pub fn new(tag: Option<String>, author: Option<String>, favorited: Option<String>, limit: Option<i64>, offset: Option<i64>) -> Self {
         Self { tag, author, favorited, limit, offset }
     }
 }
